move helpers to proper file
